Added bulk loading from the global store

+++ src/main/thrift/testclient.thrift
@@ -6,7 +6,7 @@ struct TResult {
     3: required i32 maxVal;
     4: required double dataTput;
     5: required double opsTput;
-    6: required i32 recsTput;
+    6: required double recsTput;
     7: required list<i32> tputLogKeys; //Throughput timeline log keys
     8: required list<i32> tputLogValues; //Throughput timeline log values
     9: required list<i32> latHistKeys; //Latency histogram keys
